fix: improve user interface of the language settings
